+++ book/examples/database/src/cli.rs
@@ -1,8 +1,7 @@
 use crate::app::MyApp;
 use clap::Parser;
-use roadster::api::cli::RunCommand;
+use roadster::api::cli::{CliState, RunCommand};
 use roadster::app::context::AppContext;
-use roadster::app::PreparedApp;
 use roadster::error::RoadsterResult;
 use sea_orm::prelude::async_trait::async_trait;
 
@@ -11,7 +10,7 @@ pub struct Cli;
 
 #[async_trait]
 impl RunCommand<MyApp, AppContext> for Cli {
-    async fn run(&self, _prepared_app: &PreparedApp<MyApp, AppContext>) -> RoadsterResult<bool> {
+    async fn run(&self, _prepared_app: &CliState<MyApp, AppContext>) -> RoadsterResult<bool> {
         todo!()
     }
 }
