+++ examples/app-builder/src/main.rs
@@ -1,12 +1,12 @@
 use app_builder::build_app;
-use roadster::app;
+use roadster::app::run;
 use roadster::error::RoadsterResult;
 
 #[tokio::main]
 async fn main() -> RoadsterResult<()> {
     let app = build_app();
 
-    app::run(app).await?;
+    run(app).await?;
 
     Ok(())
 }
