+++ examples/full/src/main.rs
@@ -1,10 +1,10 @@
 use full::app::App;
-use roadster::app;
+use roadster::app::run;
 use roadster::error::RoadsterResult;
 
 #[tokio::main]
 async fn main() -> RoadsterResult<()> {
-    app::run(App).await?;
+    run(App).await?;
 
     Ok(())
 }
