+++ src/api/cli/roadster/health.rs
@@ -1,7 +1,8 @@
 use crate::api::cli::roadster::RunRoadsterCommand;
+use crate::api::cli::CliState;
 use crate::api::core::health::health_check;
 use crate::app::context::AppContext;
-use crate::app::{App, PreparedApp};
+use crate::app::App;
 use crate::error::RoadsterResult;
 use async_trait::async_trait;
 use axum_core::extract::FromRef;
@@ -25,15 +26,15 @@ where
     AppContext: FromRef<S>,
     A: App<S>,
 {
-    async fn run(&self, prepared_app: &PreparedApp<A, S>) -> RoadsterResult<bool> {
+    async fn run(&self, cli: &CliState<A, S>) -> RoadsterResult<bool> {
         let duration = self
             .max_duration
             .map(Duration::from_millis)
             .unwrap_or_else(|| {
-                let context = AppContext::from_ref(&prepared_app.state);
+                let context = AppContext::from_ref(&cli.state);
                 context.config().health_check.max_duration.cli
             });
-        let health = health_check(&prepared_app.state, Some(duration)).await?;
+        let health = health_check(&cli.state, Some(duration)).await?;
         let health = serde_json::to_string_pretty(&health)?;
         info!("\n{health}");
         Ok(true)
